my_project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
numpy==2.0.1
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.1.0
spacy==3.7.5
python-dateutil==2.8.2
pydantic==2.5.0






dateparser
regex
python-dotenv



